Update oral-multidose.html
I was drinking another coffee this morning when it dawned on me that this coffee was a lot more concentrated than my first. Custom interruption dosing added. For when you drink a coffee then a monster or slam a caffeine pill I suppose. Also formatted date and time to be on top of each other to help mobile spacing. Also left/right hovering fixed.