markdown_preview init
I had this on my mind since I made a dummy link to it on the tools site while working on it to give me anxiety about someone clicking a dead link. This helped speed up production of this code. It works well enough, but as I use it for one off tasks I'll likely discover some bugs as usual.